AP_CANManager: create a new class for MAVLink CAN
