+++ libraries/AP_CANManager/AP_CANManager.h
@@ -28,8 +28,7 @@
 #include "AP_CANDriver.h"
 #include <GCS_MAVLink/GCS_config.h>
 #if HAL_GCS_ENABLED
-#include <GCS_MAVLink/GCS_MAVLink.h>
-#include <AP_HAL/utility/RingBuffer.h>
+#include "AP_MAVLinkCAN.h"
 #endif
 
 #include "AP_CAN.h"
@@ -106,9 +105,20 @@ public:
     static const struct AP_Param::GroupInfo var_info[];
 
 #if HAL_GCS_ENABLED
-    bool handle_can_forward(mavlink_channel_t chan, const mavlink_command_int_t &packet, const mavlink_message_t &msg);
-    void handle_can_frame(const mavlink_message_t &msg);
-    void handle_can_filter_modify(const mavlink_message_t &msg);
+    inline bool handle_can_forward(mavlink_channel_t chan, const mavlink_command_int_t &packet, const mavlink_message_t &msg)
+    {
+        return mavlink_can.handle_can_forward(chan, packet, msg);
+    }
+
+    inline void handle_can_frame(const mavlink_message_t &msg)
+    {
+        mavlink_can.handle_can_frame(msg);
+    }
+
+    inline void handle_can_filter_modify(const mavlink_message_t &msg)
+    {
+        mavlink_can.handle_can_filter_modify(msg);
+    }
 #endif
 
 private:
@@ -183,33 +193,8 @@ private:
     HAL_Semaphore _sem;
 
 #if HAL_GCS_ENABLED
-    /*
-      handler for CAN frames from the registered callback, sending frames
-      out as CAN_FRAME messages
-    */
-    void can_frame_callback(uint8_t bus, const AP_HAL::CANFrame &frame, AP_HAL::CANIface::CanIOFlags flags);
-
-    struct {
-        mavlink_channel_t chan;
-        uint8_t system_id;
-        uint8_t component_id;
-        uint8_t frame_counter;
-        uint32_t last_callback_enable_ms;
-        HAL_Semaphore sem;
-        uint16_t num_filter_ids;
-        uint16_t *filter_ids;
-        uint8_t callback_id;
-        uint8_t callback_bus;
-    } can_forward;
-
-    // buffer for MAVCAN frames
-    struct BufferFrame {
-        uint8_t bus;
-        AP_HAL::CANFrame frame;
-    };
-    ObjectBuffer<BufferFrame> *frame_buffer;
-
-    void process_frame_buffer(void);
+    // Class for handling MAVLink CAN frames
+    AP_MAVLinkCAN mavlink_can;
 #endif // HAL_GCS_ENABLED
 
 #if AP_CAN_LOGGING_ENABLED && HAL_LOGGING_ENABLED
